% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sampler}
\alias{sampler}
\title{sampler}
\usage{
sampler(theData, proportion = "proportation like .9")
}
\description{
Splits data into test/train dataset
}
\examples{
testTrain <- sampler(iris, .9)
testTrain$test
testTrain$train
}
